Bug 1123706 - Make proxxy use an error_level of INFO for download retries
proxxy attempts to download a file three times from each proxy URL, before
moving onto the next. There is already handling in case all proxy URLS have been
tried and the download was not successful - so it's not correct to use a log
level of ERROR for each attempt leading up to that point, since one URL may 404
but the next succeed.