Bug 1139066 part 2: Add support to download configs from git remotes.
Currently, we are able to download configs for phones only from
mercurial remotes. This is awkward because it prevents us from
testing patches in our private github repositories.

We now detect if the remote is a git or mercurial repository.

Bear in mind that git does not support downloads of individual files,
so each git provider has its own method for doing that. wE currently
support git.mozilla.org and github.com.